#moved scope groups to its package #default scope now checks if stored version is newer than requested #added get and set EntryScope extension functions for easier usage
