+++ entries/src/main/kotlin/com/source/bricks/entry/EntryFactory.kt
